+++ test/spec/format_url.js
@@ -1,4 +1,4 @@
-describe("url", function() {
+describe("format_url", function() {
     var FORCE_HTTPS;
 
     beforeEach(function() {
@@ -10,21 +10,28 @@ describe("url", function() {
     });
 
     it('returns a v4 URL with access_token parameter', function() {
-        expect(internals.url('/user.map.json')).to.equal('http://a.tiles.mapbox.com/v4/user.map.json?access_token=key')
+        expect(internals.url('/v4/user.map.json')).to.equal('http://a.tiles.mapbox.com/v4/user.map.json?access_token=key')
     });
 
     it('uses provided access token', function() {
-        expect(internals.url('/user.map.json', 'token')).to.equal('http://a.tiles.mapbox.com/v4/user.map.json?access_token=token')
+        expect(internals.url('/v4/user.map.json', 'token')).to.equal('http://a.tiles.mapbox.com/v4/user.map.json?access_token=token')
     });
 
     it('throws an error if no access token is provided', function() {
         L.mapbox.accessToken = null;
-        expect(function() { internals.url('/user.map.json') }).to.throwError('An API access token is required to use Mapbox.js.');
+        expect(function() { internals.url('/v4/user.map.json') }).to.throwError('An API access token is required to use Mapbox.js.');
     });
 
     it('throws an error if a secret access token is provided', function() {
         L.mapbox.accessToken = 'sk.abc.123';
-        expect(function() { internals.url('/user.map.json') }).to.throwError('Use a public access token (pk.*) with Mapbox.js.');
+        expect(function() { internals.url('/v4/user.map.json') }).to.throwError('Use a public access token (pk.*) with Mapbox.js.');
+    });
+
+    it('dedupes version out of custom set url', function() {
+        internals.config.FORCE_HTTPS = true;
+        internals.config.HTTPS_URL = 'https://api-maps-staging.tilestream.net/v4';
+        expect(internals.url('/v4/ludacris.map.json')).to.equal('https://api-maps-staging.tilestream.net/v4/ludacris.map.json?access_token=key');
+        internals.config.HTTPS_URL = 'https://a.tiles.mapbox.com';
     });
 
     describe('.tileJSON', function() {