+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v2.2.4
+
+* Sanitize TileJSON for `L.mapbox.shareControl`
+
+## v2.2.3
+
+* Updates to Leaflet 0.7.7.
+* Upgrades Mapbox Geocoding to v5
+
 ## v2.2.2
 
 * Updates to Leaflet 0.7.5, fixing a performance regression in tile loading